Make output buffer size user definable
A new command line option, -r, was added with which the user can specify
the size of the buffer used to store random data instead of it being a
fixed size as before. Under normal operation the buffer is filled with
random data then re-used as many times as necessary to fill the file.

The option of filling the whole file with completely random data, with
no repeats was also added.

TMain was overhauled to write data in both ways per the -r option.

If -r is not provided the same size buffer as before is used.